examples/cn0565/requirements.txt
numpy
scipy
matplotlib
pyserial
PyQt5
